rename(model): Move model (domain) files into model (domain) folder
