+++ snapnow-domain/src/main/java/kr/flab/snapnow/domain/user/domain/enums/setting/settingOption/SettingOption.java
@@ -1,0 +1,10 @@
+package kr.flab.snapnow.domain.user.domain.enums.setting.settingOption;
+
+import kr.flab.snapnow.domain.user.domain.enums.setting.SettingCategory;
+
+public interface SettingOption {
+
+    String getCode();
+    String getValue();
+    SettingCategory getCategory();
+}
