+++ snapnow-domain/src/main/java/kr/flab/snapnow/domain/user/domain/model/userSetting/PrivacySetting.java
@@ -1,11 +1,11 @@
-package kr.flab.snapnow.domain.user.model.userSetting;
+package kr.flab.snapnow.domain.user.domain.model.userSetting;
 
 import java.util.HashMap;
 
-import kr.flab.snapnow.domain.user.enums.setting.SettingCategory;
-import kr.flab.snapnow.domain.user.enums.setting.settingOption.PrivacySettingOption;
-import kr.flab.snapnow.domain.user.enums.setting.settingType.PrivacySettingType;
-import kr.flab.snapnow.domain.user.exception.InvalidSettingException;
+import kr.flab.snapnow.domain.user.domain.enums.setting.SettingCategory;
+import kr.flab.snapnow.domain.user.domain.enums.setting.settingOption.PrivacySettingOption;
+import kr.flab.snapnow.domain.user.domain.enums.setting.settingType.PrivacySettingType;
+import kr.flab.snapnow.domain.user.domain.exception.InvalidSettingException;
 
 public class PrivacySetting extends UserSetting<PrivacySettingType, PrivacySettingOption> {
