--- snapnow-domain/src/main/java/kr/flab/snapnow/domain/user/enums/setting/settingOption/SettingOption.java
@@ -1,10 +1,0 @@
-package kr.flab.snapnow.domain.user.enums.setting.settingOption;
-
-import kr.flab.snapnow.domain.user.enums.setting.SettingCategory;
-
-public interface SettingOption {
-
-    String getCode();
-    String getValue();
-    SettingCategory getCategory();
-}
